Tarefa1

Description

Introdução

Nesta tarefa, tinhamos como objetivo implementar um mecanismo de geração de labirintos.

Metas

A principal meta desta tarefa era conseguir gerar um labirinto, onde os inputs seriam: a sua largura, a sua altura e uma seed (um número inteiro), respetivamente. Dito isto, neste labirinto tinhamos que incluir um túnel, que variava de tamanho face ao número par (túnel formado por 2 corredores) ou ímpar (túnel formado por 1 corredor) de altura. Para além do túnel, tinhamos que também desenvolver a casa dos fantasmas cuja altura fixa era 3 e variava de largura (largura 9 se a largura do labirinto fosse ímpar ou 8 se fosse par). Isto, para além de todas as outras funções que tivemos de desenvolver pedidas no trabalho e/ou auxiliadoras no processo de geração do labirinto.

Conclusão

Em suma, penso que apesar de algumas dificuldades iniciais, fizemos um bom trabalho face aos nossos conhecimentos atuais. Penso que conseguimos atingir o nosso principal objetivo desta tarefa que era passar em todos os testes no sistema de avaliação dos professores, mesmo sabendo que podíamos melhorar em alguns aspetos, tal como, a qualidade do código.

Documentation

geraAleatorios :: Int -> Int -> [Int]

Dada uma seed devolve uma lista de n inteiros gerados aleatóriamente.

nrAleatorio :: Int -> Int

Dada uma seed devolve um inteiro gerado aleatóriamente.

subLista :: Int -> [a] -> [[a]]

Converte uma lista em uma lista de listas de tamanho n.

convertPiece :: Int -> Piece

Converte número numa peça.

printCorridor :: Corridor -> String

Converte o corredor numa string.

printMaze :: Maze -> String

Converte o labirinto numa string.

convertCorridor :: [Int] -> Corridor

Converte uma lista de números interios em corredor chamando a função convertePiece.

convertMaze :: [[Int]] -> Maze

Converte uma lista de listas de números inteiros em labirinto chamando as funções replaceWithWall e convertCorridor.

genWall :: Int -> Corridor

Gera corredores do tipo Parede (sempre horizontal).

putWall :: Int -> Maze -> Maze

Transforma o topo e o fundo do labirinto em corredores do tipo Parede.

replaceWithWall :: Corridor -> Corridor

Substitui primeira e última peça de cada corredor com Wall.

middleHeight :: Maze -> Int

Calcula metade da altura do labirinto.

middleWidth :: Maze -> Int

Calcula metade da largura do labirinto.

middleCorridors :: Maze -> [Corridor]

Obtém os corredores do meio do labirinto.

breakWall :: [Corridor] -> [Corridor]

Acrescenta peça Empty para criar entrada no(s) corredor(es) do túnel.

openTunnel :: Maze -> Maze

Constrói o túnel chamando as funções middleHeight, breakWall e middleCorridors.

buildHouse :: Maze -> Maze

Constrói casa dos fantasmas chamando as funções middleHeight e middleCorridorsh.

house :: [Corridor] -> [Corridor]

Formato da casa dos fantasmas.

middleCorridorsh :: Maze -> [Corridor]

Obtém os corredores do meio (função auxiliar para criação da casa dos fantasmas).

generateMaze :: Int -> Int -> Int -> Maze

Gera o labirinto.