Tarefa2

Description

Introdução

Nesta tarefa, tinhamos como objetivo determinar o efeito de uma jogada no estado de jogo (dada uma descrição do estado do jogo e uma jogada de um dos jogadores).

Metas

Tal como mencionado na introdução, a meta desta tarefa era determinar o efeito de uma jogada no estado do jogo (sabendo que esta tarefa apenas exigia a implementação de realização de jogadas para jogadores do tipo Pacman). Cada jogada consiste na movimentação de um jogador identificado com o seu único ID (para todas as direções possíveis, isto é, cima, baixo, direita, esquerda) e atualizar o efeito dessa jogada no labirinto atual (função objetivo da tarefa play). Para tal, tivemos de criar as funções que de certo modo, traduziam as reações do jogadores e/ou do labirinto face a todas as possíveis situações, criando deste modo todos os possíveis resultados.

Conclusão

Concluindo este resumo da tarefa 2, afirmamos que, esta foi a tarefa mais trabalhosa da primeira parte do trabalho prático, que no entanto, apesar de várias alterações no código, de forma a que a tarefa fosse completa corretamente e, eventualmente, de maneira a que a qualidade do código fosse melhorada, pensamos que o nosso trabalho foi bastante satisfatório, passando em todos os testes no sistema de avaliação dos professores.

Documentation

play :: Play -> State -> State

Faz uma jogada sobre um determinado jogador e devolve um novo estado do jogo.

freezePlayers :: [Player] -> [Player]

Congela todos os jogadores.

updateMaze :: Player -> Maze -> Maze

Altera o labirinto.

gameWithMega :: [Player] -> [Player]

Consoante o estado do jogo onde haja ou não algum jogador Pacman em modo Mega, altera o estado dos fantasmas.

checkPlayers :: Maze -> [Player] -> [Player]

Compara um jogador com vários e altera o seu estado caso seja ncessário.

playersWithSameCoords :: Maze -> Player -> [Player] -> Player

Verifica se tem jogadores na mesma coordenada e devolve o jogador com estado alterado.

actionPlayerPlayer :: Maze -> Player -> Player -> Player

Muda o estado de um jogador consoante o jogador em que embate.

teleportToHouse :: Maze -> Player -> Player

Teletransporta um fantasma para a casa de fantasmas.

joinHouse :: Maze -> Coords -> Coords -> Bool

Verifica se o jogador vai entrar na casa.

playerMaze :: Play -> [Player] -> Maze -> [Player]

Relação jogador & labirinto.

checkNextPiece :: Coords -> Maze -> Piece

Procura no labirinto a peça na posição para a qual um jogador transita.

checkPiece :: Int -> Corridor -> Piece

Devolve a peça que se encontra na posição para a qual um jogador transita.

actionPlayerPiece :: Player -> Piece -> Player

Muda o estado de um jogador consoante a peça em que o jogador embate.

pacmanDefaultMegaTime :: Double

Tempo em modo Mega do Pacman.

setPlayerOrientation :: Player -> Orientation -> Player

Define as coordenadas de um jogador.

getPlayerCoordsByID :: Int -> [Player] -> Coords

Devolve as coordenadas de um determinado jogador.

verifyNewMega :: [Player] -> Bool

Verifica se um jogador Pacman entrou em modo Mega.

verifyNoMega :: [Player] -> Bool

Verifica se não existe um jogador Pacman em modo Mega.

turnGhostsAlive :: [Player] -> [Player]

Muda o estado dos fantasmas para Alive.

turnGhostsDead :: [Player] -> [Player]

Muda o estado dos fantasmas para Dead.

invertOrientation :: Orientation -> Orientation

Inverte a oritenção.

getNewCoords :: Coords -> Orientation -> Coords

Devolve as novas coordenadas de um jogador consoante a sua orientação.

tunnelTeleport :: Player -> Int -> Player

Teletransporta um jogador que entrou no túnel.

middleHeight :: Maze -> Int

Calcula metade da altura do labirinto

middleWidth :: Maze -> Int

Calcula metade da largura do labirinto

mazeWidth :: Maze -> Int

Calcula a largura total do labirinto