Tarefa4

Documentation

passTime :: Int -> State -> State

Introdução

Nesta tarefa, tinhamos como objetivo calcular o efeito da passagem de um instante de tempo num estado do jogo.

Metas

Tal como mencionado na introdução, a meta desta tarefa era calcular o efeito da passagem de um instante de tempo num estado de jogo. Dito isto, a tarefa estava dividida em 2 componentes, a primeira onde integraríamos os tipos e funções que advêm da 1ª fase numa biblioteca denominada Ncurses, a segunda que usa o estado e informação vinda do primeiro componente para fazer avançar o estado em uma iteração, fazendo “movimentar” todos os jogadores no mapa. De certa forma, as principais metas desta tarefa consistiam na integração da biblioteca Ncurses, na construção de algumas funções que alteram o estado dos jogadores e alteração de outras (de tarefas anteriores), bem como funções para execução de certas jogadas e, finalmente, funções que trabalhavam a velocidade dos jogadores em função da paridade da iteração (step). Deste modo, conseguimos chegar à principal meta/função que tem como objetivo a execução de jogadas consoante o tempo de jogo.

Conclusão

Concluindo este resumo da Tarefa4, para obter um resultado satisfatório nesta tarefa, necessitámos de reformular a maioria do código das funções da Tarefa2, de maneira a melhorar a performance do jogo, bem como coexistir com todas as novas funções adicionadas na Tarefa4, de forma a que fosse possível calcular o efeito da passagem de tempo no estado de jogo. Em suma, julgo que conseguimos obter resultados produtivos, apesar de poder melhor em alguns aspetos, tal como qualidade de código, entre outros.

Executa jogadas consoante o tempo de jogo.

pacmanMouth :: Player -> Player

Muda o estado da boca do Pacman (Aberta/Fechada) a cada jogada.

pacmanMegaTime :: Player -> Player

Altera o tempo restante em modo Mega do Pacman.

movePlayers :: Int -> State -> [Player] -> State

Movimenta todos os jogadores no labirinto.

movePlayer :: Int -> Play -> State -> State

Move o jogador consoante a sua velocidade.

getPlayerVelocity :: Player -> Double

Devolve a velocidade do jogador.

playerMoves :: Int -> Double -> Int

Devolve quantas vezes o jogador vai jogar numa determinada iteração.

editPlayers :: State -> State

Altera estado de alguns jogadores.

editPlayersAux :: [Player] -> [Player]

Altera estado de alguns jogadores.

executeGhostPlay :: [Play] -> State -> State

Executa as jogadas dos fantasmas (bots).

beforePlay :: Int -> Orientation -> State -> State

Executa uma jogada apenas se o jogador alterar a orientação.