Tarefa5

Documentation

ghostPlay :: State -> [Play]

Introdução

Nesta tarefa, tinhamos como objetivo implementar um comportamento para os fantasmas.

Metas

Tal como mencionado na introdução, a meta desta tarefa era implementar um comportamento para os fantasmas. Para chegar a esta meta, inicialmente, tivemos que complementar novamente, as tarefas da Tarefa2, já que necessitávamos de considerar que a função play poderia receber jogadas de fantasmas. Para além disso, desenvolvemos funções que complementavam a movimentação do fantasma. Estas funções são diferenciadas pelo estado do fantasma, isto é, quando o fantasma está Alive, ele procura encontrar o Pacman ( de formas variadas como movimentação na horizontal e na vertical ou aleatória), por outro lado, quando está Dead, ele procura fugir do Pacman. Deste modo conseguimos implementar um comportamento para os fantasmas.

Conclusão

Concluindo este resumo da Tarefa5, para obter um resultado satisfatório, tivemos que complementar algumas funções de tarefas anteriores (nomeadamente Tarefa2 e Tarefa4), de forma a conseguir desenvolver um conjunto de jogadas diferenciadas para os fantasmas. Em suma, penso que o trabalho desenvolvido pelo grupo nesta tarefa foi bastante razoável.

Devolve um conjunto de jogadas feitas pelos bots.

leaveHouse :: State -> [Player] -> [Play]

Tira o jogador da casa dos fantasmas.

chaseMode :: State -> Int -> Play

Movimenta os fantasmas para caçarem o Pacman.

scatterMode :: State -> Int -> Play

Movimenta os fantasmas em modo Dead para fugirem do Pacman.

horizontalMove :: Int -> [Player] -> Play

Movimenta o fantasma com prioridade na horizontal.

verticalMove :: Int -> [Player] -> Play

Movimenta o fantasma com prioridade na vertical.

ghostHorizontal :: Coords -> Coords -> Orientation

Altera a orientação do fantasma com prioridade horizontal.

ghostVertical :: Coords -> Coords -> Orientation

Altera a orientação do fantasma com prioridade vertical.

comparePlayerState :: Player -> Player -> Bool

Compara o estado do fantasma em dois estados de jogo diferentes.

goRight :: Orientation -> Orientation

Altera a orientação do fantasma no sentido dos ponteiros do relógio.

goLeft :: Orientation -> Orientation

Altera a orientação do fantasma no sentido contrário aos ponteiros do relógio.

ghostCollideWall1 :: State -> Play -> Play

Verifica se o fantasma (Alive) colidiu em uma parede e altera a orientação.

ghostCollideWall2 :: State -> Play -> Play

ghostCollideWall3 :: State -> Play -> Play

ghostCollideWall4 :: State -> Play -> Play

ghostScatterCollideWall :: State -> Play -> Play

Verifica se o fantasma (Dead) colidiu em uma parede e altera a orientação.