Tarefa6

Description

Introdução

Nesta tarefa, tinhamos como objetivo implementar um robô que jogue Pacman normalmente.

Metas

Tal como mencionado na introdução, a meta desta tarefa era implementar um robô que jogue Pacman normalmente. Para chegar a esta meta, necessitámos de criar funções para movimentação do bot. Para além da movimentação do robô tentámos fazer com que este verificasse se na sua proximidade havia a potencial existência de um fantasma, tentando desviar-se deste e podendo, movimentar-se noutra direção onde existissem comidas.

Conclusão

Concluindo este resumo da Tarefa6, na nossa sincera opinião, foi a tarefa onde encontramos mais dificuldades, mas julgamos que obtivemos um resultado mais ao menos satisfátorio, apesar de poder melhorar em alguns aspetos, tal como a criação de ainda mais "niveis" e/ou prioridades para o bot, de maneira a que este ficasse ainda mais "desenvolvido".

Documentation

bot :: Int -> State -> Maybe Play

Devolve uma possível jogada do bot.

checkDirection :: Orientation -> Coords -> [Player] -> Bool

Verifica se na direção atual irá encontrar algum fantasma num espaço de 3 casas.

checkGhostsCoords :: Int -> Orientation -> Coords -> Player -> Bool

Verifica se um fantasma está numa determinada posição.

getCorridor :: Int -> Maze -> Corridor

Devolve o corredor em que o jogador está.

checkFood :: Corridor -> Bool

Verifica se exite alguma comida num dado corredor.

changeBotOrientation :: Orientation -> Orientation

Muda a orientação do bot.

comparePacState :: Player -> Player -> Bool

Compara o estado do fantasma em dois estados de jogo diferentes.

botCollideWall :: State -> Play -> Play

Verifica se bot colidiu em uma parede e altera a orientação.